+++ framework/core/plugins/signalR/index.ts
@@ -8,9 +8,8 @@ import { useCypressSignalRMock } from "cypress-signalr-mock";
 const { addNotification } = useNotifications();
 const currentCreator = ref<string | undefined>();
 
-export const updateSignalRCreatorSymbol = Symbol("updateSignalRCreator") as InjectionKey<
-  (creator: string | undefined) => void
->;
+export const updateSignalRCreatorSymbol: InjectionKey<(creator: string | undefined) => void> =
+  Symbol("updateSignalRCreator");
 
 function setupSystemEventsHandler(connection: HubConnection, creator?: string) {
   // Unsubscribe from the previous handler if it was