+++ framework/shared/components/sidebar/sidebar.vue
@@ -30,6 +30,8 @@
       <slot name="content"></slot>
     </div>
   </template>
+
+  <!-- Pass through content if not expanded -->
   <template v-else>
     <slot name="content"></slot>
   </template>