+++ framework/shared/modules/dynamic/pages/dynamic-blade-list.vue
@@ -750,6 +750,7 @@ const tableConfig = computed(() => {
     "onSelect:all": handleSelectAllItems,
     onEditComplete: onEditComplete,
     onCellBlur: onCellBlur,
+    noHeaderCheckbox: tableData?.value?.noHeaderCheckbox,
   };
 });