+++ framework/ui/components/organisms/vc-table/vc-table.vue
@@ -4,7 +4,7 @@
     class="vc-table"
   >
     <div
-      v-if="multiselect && $isMobile.value && (selection.length > 0 || allSelected)"
+      v-if="multiselect && $isMobile.value && (selection.length > 0 || allSelected) && !noHeaderCheckbox"
       class="vc-table__multiselect-mobile"
     >
       <div class="vc-table__select-all-bar">
@@ -156,6 +156,7 @@
               >
                 <div class="vc-table__header-checkbox__content">
                   <VcCheckbox
+                    v-if="!noHeaderCheckbox"
                     v-model="headerCheckbox"
                     size="m"
                     @click.stop
@@ -550,6 +551,7 @@ const props = withDefaults(
     selectionItems?: T[];
     disableFilter?: boolean;
     columnSelector?: "auto" | "defined" | MaybeRef<ITableColumns[]> | (() => ITableColumns[]);
+    noHeaderCheckbox?: boolean;
   }>(),
   {
     items: () => [],
@@ -612,7 +614,7 @@ const table = useCurrentElement();
 const resizer = ref();
 const isHeaderHover = ref(false);
 const columnSwitcherActive = ref(false);
-const state = useLocalStorage<TableColPartial[]>("VC_TABLE_STATE_" + props.stateKey.toUpperCase(), []);
+const state = useLocalStorage<TableColPartial[]>("VC_TABLE_STATE_" + props.stateKey?.toUpperCase(), []);
 const internalColumns: Ref<TableColPartial[]> = ref([]);
 const draggedColumn = ref();
 const draggedElement = ref<HTMLElement>();